Add files via upload
Uses a kd-tree to find the p number of near neighbours for each point in an input/output dataset. The advantage of the kd-tree is that it runs in O(M log M) time. In this project, it helped to find the nearest bike station ID for every taxi ride in New York city with the help of rdgal and RANN packages in R